Split out counts and start adding haemonc cancer counts
