+++ merge_sample_info.py
@@ -1,7 +1,7 @@
 import argparse
 import pandas as pd
 
-from utils import read_in_to_df
+from utils.file_io import read_in_to_df
 
 
 def parse_args() -> argparse.Namespace:
